[profiling] Refactor: use the new FFI macros and types
